API/requirements.txt
django==2.1.3
djangorestframework==3.8.2
django-debug-toolbar==1.10.0
django-filter==2.0.0
psycopg2-binary==2.7.6.1
jsonschema==2.6.0
incremental
gunicorn==19.9.0
requests==2.20.1
redis==3.0.1
influxdb==5.2.0
psutil==5.4.8
django-cors-headers==2.4.0
netaddr==0.7.19
fqdn==1.1.0
pamela==0.3.0
mongoengine==0.16.2
django-rest-framework-mongoengine==3.3.1
asyncio-redis==0.15.1
motor==2.0.0
pillow==5.3.0
certifi==2018.11.29
chardet==3.0.4
idna==2.7
urllib3==1.24.1
pbr==5.1.1
requests-unixsocket==0.1.5
channels==2.2.0
Twisted[tls,http2]==18.9.0
channels_redis==2.2.1
unittest-xml-reporting==2.2.0
terminaltables==3.1.0
tqdm==4.33.0
termcolor==1.1.0
django-extensions==2.2.1
